slight dsp cleanup

+++ include/JSystem/JAudio2/DSP.h
@@ -30,18 +30,18 @@ typedef struct _STRUCT_DSP_TASK {
 
 // dspproc.c
 void DSPReleaseHalt2(u32 msg);
-void setup_callback(unsigned short p1);
-void DsetupTable(unsigned long p1, unsigned long p2, unsigned long p3, unsigned long p4, unsigned long p5);
+void setup_callback(u16 p1);
+void DsetupTable(u32 p1, u32 p2, u32 p3, u32 p4, u32 p5);
 void DsetMixerLevel(f32 mixerLevel);
-void DsyncFrame(unsigned long p1, unsigned long p2, unsigned long p3);
+void DsyncFrame(u32 p1, u32 p2, u32 p3);
 
 // dsptask.c
 void DspHandShake(void* a1);
 void DspBoot(void (*p1)(void*));
-void DSPSendCommands2(unsigned long* p1, unsigned long p2, void (*p3)(unsigned short));
+void DSPSendCommands2(u32* p1, u32 p2, void (*p3)(u16));
 void DspInitWork();
-void DspStartWork(unsigned long p1, void (*p2)(unsigned short));
-void DspFinishWork(unsigned short p1);
+void DspStartWork(u32 p1, void (*p2)(u16));
+void DspFinishWork(u16 p1);
 
 // osdsp.c
 #ifdef __cplusplus
@@ -57,7 +57,7 @@ extern "C" {
 void __DSPHandler();
 }
 #endif // ifdef __cplusplus
-void DsyncFrame2(unsigned long p1, unsigned long p2, unsigned long p3);
+void DsyncFrame2(u32 p1, u32 p2, u32 p3);
 void Dsp_Update_Request();
 BOOL Dsp_Running_Check();
 void DSP_Running_Start();
@@ -68,17 +68,17 @@ void DSPSendCommands(u32* p1, u32 p2);
 void DSPReleaseHalt();
 void DSPWaitFinish();
 void Dswap(u32 p1, u32 p2, u32 p3);
-void Dmix(unsigned long p1, unsigned long p2, unsigned long p3, short p4);
-void Dcopy(unsigned long p1, unsigned long p2, unsigned long p3);
-void DloadBuffer1(unsigned long p1, unsigned long p2);
+void Dmix(u32 p1, u32 p2, u32 p3, short p4);
+void Dcopy(u32 p1, u32 p2, u32 p3);
+void DloadBuffer1(u32 p1, u32 p2);
 void DloadBuffer(u32 p1, u32 p2, u32 p3);
-void DsaveBuffer(unsigned short p1, unsigned long p2, unsigned long p3);
-void wait_callback(unsigned short p1);
+void DsaveBuffer(u16 p1, u32 p2, u32 p3);
+void wait_callback(u16 p1);
 void DwaitFrame();
-void DiplSec(unsigned long p1, void (*p2)(unsigned short));
-void DagbSec(unsigned long p1, void (*p2)(unsigned short));
-void dummy_callback(unsigned short p1);
-void DsetDolbyDelay(unsigned long p1, unsigned short p2);
+void DiplSec(u32 p1, void (*p2)(u16));
+void DagbSec(u32 p1, void (*p2)(u16));
+void dummy_callback(u16 p1);
+void DsetDolbyDelay(u32 p1, u16 p2);
 
 extern u8 DSP_prior_yield;
 extern STRUCT_DSP_TASK* DSP_prior_task;
